Restructure and cleanup parsers and encoders
