+++ ical/todo.py
@@ -7,9 +7,8 @@ from typing import Any, Optional, Union
 from pydantic import Field, validator
 
 from .contentlines import ParsedProperty
-from .model import ComponentModel
 from .properties import Priority, TodoStatus
-from .property_values import Date, DateTime, Text
+from .types import ComponentModel, Date, DateTime, Text
 
 
 class Todo(ComponentModel):
