+++ tests/test_types.py
@@ -1,4 +1,4 @@
-"""Tests for encoders."""
+"""Tests for property values."""
 
 import datetime
 import json
@@ -7,8 +7,30 @@ from typing import Optional
 from pydantic import BaseModel
 
 from ical.contentlines import ParsedComponent, ParsedProperty
-from ical.encoders import encode_component
-from ical.property_values import DateTime
+from ical.types import ComponentModel, DateTime, Text, encode_component
+
+
+class TextModel(ComponentModel):
+    """Model with a Text value."""
+
+    text_value: Text
+
+
+def test_text() -> None:
+    """Test for a text property value."""
+    component = ParsedComponent(name="text-model")
+    component.properties.append(
+        ParsedProperty(
+            name="text_value",
+            value="Project XYZ Final Review\\nConference Room - 3B\\nCome Prepared.",
+        )
+    )
+    model = TextModel.parse_obj(component.as_dict())
+    assert model == {
+        "text_value": "\n".join(
+            ["Project XYZ Final Review", "Conference Room - 3B", "Come Prepared."]
+        )
+    }
 
 
 class OtherComponent(BaseModel):
@@ -18,7 +40,7 @@ class OtherComponent(BaseModel):
     second_value: Optional[str] = None
 
 
-class TextModel(BaseModel):
+class TestModel(BaseModel):
     """Model with a Text value."""
 
     text_value: str
@@ -28,7 +50,7 @@ class TextModel(BaseModel):
     dt: DateTime
 
     class Config:
-        """Configuration for TextModel pydantic model."""
+        """Configuration for TestModel pydantic model."""
 
         json_encoders = {
             datetime.datetime: DateTime.ics,
@@ -37,7 +59,7 @@ class TextModel(BaseModel):
 
 def test_encode_component() -> None:
     """Test for a text property value."""
-    model = TextModel.parse_obj(
+    model = TestModel.parse_obj(
         {
             "text_value": "Example text",
             "repeated_text_value": ["a", "b", "c"],
@@ -52,9 +74,9 @@ def test_encode_component() -> None:
         }
     )
     component = encode_component(
-        "TextModel", json.loads(model.json(exclude_unset=True))
+        "TestModel", json.loads(model.json(exclude_unset=True))
     )
-    assert component.name == "TextModel"
+    assert component.name == "TestModel"
     assert component.properties == [
         ParsedProperty(name="text_value", value="Example text"),
         ParsedProperty(name="repeated_text_value", value="a"),
